[login hook] Map keys as login user
The `login-hook` shell script now checks whether it's being run as
a login hook (as root with $1 set) or directly (with $1 not set) so that
keys are mapped as the login user in either case. Previously, the keys
would be mapped as `root` on login since the login script runs as `root`
on login, but that doesn't seem to work--i.e., the keys weren't actually
being mapped on login, requiring `login-hook` to be run manually.