[setup, macos] [WIP] Add script to macOS preferences
This replaces the `login-hook` script for setting modifier keys and also
sets a bunch of other preferences.

The impetus for this is to swap capslock and control in the "proper"
way. The method used in `login-hook`[1] seemed to be overwritten by the
"official" System Preferences > Modifiers Keys settings such that the
`login-hook` would have to be re-run sporadically (e.g., after sleep).

The main problems with this new script is that 1) it's kind of a pain to
figure out how to set system preferences using `defaults write` (and
some settings aren't set via `defaults write`) and 2) some preferences
when set via `defaults write` aren't reflected in the macOS GUI (i.e.,
in the System Preferences pane).

Ultimately, though (hopefully), this should make setting up a new Mac a
bit less tedious.

[1] `hidutil property --set '{"UserKeyMapping": ...}`