[live-backup] Add env file option (--env-file/-f)
This allows the live backup env to be specified in a non-versioned file.
The default when not specified is to look for `.live-backup.env` in the
config file's directory.

If an env is specified with `--env` or any other way, that env will take
precedence over one specified via env file.