[live-backup] Skip missing paths by default
When there are missing project paths, the default behavior now will be
to prompt to continue rather than abort. This is a mitigation for the
fact that different machines might have different projects checked out.

In particular, I don't want to have to edit the local live-backup config
to remove paths to work around this, because it's a pain and causes
unnecessary git diffs.

With this change, hopefully the live-backup config file can be
standardized across machines.

Added `--no-skip-missing-paths` flag; when passed, missing paths will
cause the script to abort like before.